requirements.txt
anyio==3.7.0
async-generator==1.10
attrs==23.1.0
certifi==2023.5.7
click==8.1.3
exceptiongroup==1.1.1
fastapi==0.96.1
gunicorn==20.1.0
h11==0.14.0
idna==3.4
mysql-connector==2.2.9
outcome==1.2.0
pydantic==1.10.9
PySocks==1.7.1
selenium==4.9.1
sniffio==1.3.0
sortedcontainers==2.4.0
starlette==0.27.0
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.6.3
urllib3==2.0.2
uvicorn==0.22.0
wsproto==1.2.0
